% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_heatmaps.r
\name{generate_heatmaps}
\alias{generate_heatmaps}
\title{generate_heatmaps}
\usage{
generate_heatmaps(input_dir, counts_file_name, output_dir, downsample_label)
}
\arguments{
\item{input_dir}{...}

\item{counts_file_name}{...}

\item{output_dir}{...}

\item{downsample_label}{...}
}
\description{
Generate two heatmaps: one static and one interactive
}
\author{
Yuanyuan Wang
}
